cleaned up some deprecated functions ; changed some names
