+++ sql_connectors/facade_builder.py
@@ -102,7 +102,7 @@ def full_path(sub_path="", config_dir=None, default_config_dir=DEFAULT_CONFIG_DI
     return os.path.join(os.path.expanduser(config_base_dir), sub_path)
 
 
-def get_available_configs(fpath):
+def get_available_configs(fpath=None):
     """Gets all available json configs in a certain path
     
     Parameters
@@ -115,6 +115,8 @@ def get_available_configs(fpath):
     list of str
         list of json full file paths
     """
+    if fpath is None:
+        fpath = full_path()
     files = glob(os.path.join(fpath, "*.json"))
     return files
 
@@ -340,55 +342,3 @@ class BuildFacade(object):
         kwargs = self.get_factories()
         instance = Namespace(**kwargs)
         return instance
-
-
-def get_fname(fpath):
-    """Gets filename without extension from full path
-    
-    Parameters
-    ----------
-    fpath : str
-        full path
-    
-    Returns
-    -------
-    str
-    """
-    return fpath.split("/")[-1].replace(".json", "")
-
-
-def get_default_params(config_path=None):
-    """Gets default params from a config location on disk
-    
-    Parameters
-    ----------
-    config_path : str, optional
-        config location on disk - defaults to the SQL_CONNECTORS_CONFIG_DIR environment var
-    
-    Returns
-    -------
-    tuple of dicts
-        iterable of config file names to pass into `BuildFacade`
-    """
-    configs = glob(full_path("*json", config_dir = config_path))
-    names = [get_fname(i) for i in configs]
-    return tuple({"name": name} for name in names)
-
-
-def default_facade(**kwargs):
-    """Function to build out the connector facade from the default config path
-
-    Parameters
-    ----------
-    **kwargs
-    config_path: str
-        location of config file - defaults to the SQL_CONNECTORS_CONFIG_DIR environment var
-
-    Returns
-    -------
-    argparse.Namespace
-    """
-    config_path = kwargs.get("config_path", None)
-    params = get_default_params(config_path)
-    cls = BuildFacade(params)
-    return cls.build_factory_facade()
